Added test "groups".
Typically in jUnit or other common OO unit test frameworks, tests are
conceived of as methods of classes which are iterated
over. Unfortunately, this pattern simply does not apply in Common Lisp
very well, as methods are free and unattached to classes.

Taking fiveam's lead, I create mechanism to "group" tests together in a
simple fashion and provide an API for controlling which group a test is
designated to live under.